refactored code generation
