+++ oximod_macros/src/lib.rs
@@ -6,9 +6,9 @@ use std::collections::HashSet;
 use proc_macro::TokenStream;
 use quote::quote;
 use syn::{ parse_macro_input, DeriveInput, LitStr };
-use index::parse_index_args;
-use validate::{ parse_validate_args, ValidateArgs };
-use default::{ parse_default_args, option_inner_type };
+use index::{ parse_index_args, generate_index_model_tokens };
+use validate::{ parse_validate_args, generate_validate_model_tokens };
+use default::{ parse_default_args, maybe_push_id_setter, push_field_setters };
 
 #[proc_macro_derive(
     Model,
@@ -127,269 +127,13 @@ pub fn derive_model(input: TokenStream) -> TokenStream {
         }
     }
 
-    let index_models = index_definitions.iter().map(|index_def| {
-        let field = &index_def.field_name;
-        let order = index_def.args.order.unwrap_or(1);
+    let index_models = index_definitions
+        .iter()
+        .map(|index_def| generate_index_model_tokens(index_def));
 
-        let unique = match index_def.args.unique {
-            Some(val) => quote! { Some(#val) },
-            None => quote! { None },
-        };
-
-        let sparse = match index_def.args.sparse {
-            Some(val) => quote! { Some(#val) },
-            None => quote! { None },
-        };
-
-        let background = match index_def.args.background {
-            Some(val) => quote! { Some(#val) },
-            None => quote! { None },
-        };
-
-        let name = match &index_def.args.name {
-            Some(val) => quote! { Some(#val.to_string()) },
-            None => quote! { None },
-        };
-
-        let expire_after_secs = match index_def.args.expire_after_secs {
-            Some(secs) => {
-                quote! { Some(::std::time::Duration::from_secs(#secs as u64)) }
-            }
-            None => quote! { None },
-        };
-
-        quote! {
-            ::oximod::_mongodb::IndexModel::builder()
-                .keys(::oximod::_mongodb::bson::doc! { #field: #order })
-                .options(
-                    ::oximod::_mongodb::options::IndexOptions::builder()
-                        .unique(#unique)
-                        .sparse(#sparse)
-                        .background(#background)
-                        .name(#name)
-                        .expire_after(#expire_after_secs)
-                        .build()
-                )
-                .build()
-        }
-    });
-
-    let validations = validate_definitions.iter().flat_map(|validate_def| {
-        let field_ident = syn::Ident::new(&validate_def.field_name, proc_macro2::Span::call_site());
-        let ValidateArgs {
-            min_length,
-            max_length,
-            required,
-            // enum_values,
-            email,
-            pattern,
-            non_empty,
-            positive,
-            negative,
-            non_negative,
-            min,
-            max,
-        } = &validate_def.args;
-
-        let mut checks = vec![];
-
-        if let Some(min) = min_length {
-            checks.push(
-                quote! {
-                    if self.#field_ident.len() < #min as usize {
-                        return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                            format!("Field '{}' must be at least {} characters long", stringify!(#field_ident), #min)
-                        ));
-                    }
-                }
-            );
-        }
-
-        if let Some(max) = max_length {
-            checks.push(
-                quote! {
-                    if self.#field_ident.len() > #max as usize {
-                        return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                            format!("Field '{}' must be at most {} characters long", stringify!(#field_ident), #max)
-                        ));
-                    }
-                }
-            );
-        }
-
-        if let Some(req) = required {
-            if *req {
-                checks.push(
-                    quote! {
-                        match self.#field_ident {
-                            Some(_) => {},
-                            None => {
-                                return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                                    format!("Field '{}' is required", stringify!(#field_ident))
-                                ))
-                            },
-                            _ => {}
-                        }
-                    }
-                );
-            }
-        }
-
-        // if let Some(values) = enum_values {
-        //     let allowed: Vec<proc_macro2::TokenStream> = values
-        //         .iter()
-        //         .map(|v| quote! { #v })
-        //         .collect();
-
-        //     checks.push(
-        //         quote! {
-        //             if let Some(ref value) = self.#field_ident {
-        //                 if ! [#( #allowed ),*].contains(&value.as_str()) {
-        //                     return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-        //                         format!(
-        //                             "Field '{}' must be one of: {}",
-        //                             stringify!(#field_ident),
-        //                             vec![#( #allowed.to_string() ),*].join(", ")
-        //                         )
-        //                     ));
-        //                 }
-        //             }
-        //         }
-        //     );
-        // }
-
-        if let Some(is_email) = email {
-            if *is_email {
-                checks.push(
-                    quote! {
-                        if let Some(email) = &self.#field_ident {
-                            if !email.contains('@') || !email.contains('.') {
-                                return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                                    format!("Field '{}' must be a valid email address", stringify!(#field_ident))
-                                ));
-                            }
-                        
-                            let parts: Vec<&str> = email.split('@').collect();
-                            if parts.len() != 2 || parts[0].is_empty() || parts[1].is_empty() || !parts[1].contains('.') {
-                                return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                                    format!("Field '{}' must be a valid email address", stringify!(#field_ident))
-                                ));
-                            }
-                        } 
-                    }
-                );
-            }
-        }
-
-        if let Some(pattern) = pattern {
-            checks.push(
-                quote! {
-                if let Some(ref value) = self.#field_ident {
-                    let regex = ::oximod::_regex::Regex::new(#pattern).map_err(|e| {
-                        ::oximod::_error::oximod_error::OximodError::ValidationError(
-                            format!("Invalid regex pattern in validation for '{}': {}", stringify!(#field_ident), e)
-                        )
-                    })?;
-                    if !regex.is_match(value) {
-                        return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                            format!(
-                                "Field '{}' does not match the required pattern",
-                                stringify!(#field_ident)
-                            )
-                        ));
-                    }
-                }
-            }
-            );
-        }
-
-        if let Some(true) = non_empty {
-            checks.push(
-                quote! {
-                let value = &self.#field_ident;
-                if let Some(ref val) = value {
-                    if val.trim().is_empty() {
-                        return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                            format!("Field '{}' must be non-empty", stringify!(#field_ident))
-                        ));
-                    }
-                } else {
-                    return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                        format!("Field '{}' is missing but marked as non-empty", stringify!(#field_ident))
-                    ));
-                }
-            }
-            );
-        }
-
-        if let Some(positive) = positive {
-            if *positive {
-                checks.push(
-                    quote! {
-                        if self.#field_ident <= 0 {
-                            return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                                format!("Field '{}' must be positive", stringify!(#field_ident))
-                            ))
-                        }
-                    }
-                );
-            }
-        }
-
-        if let Some(negative) = negative {
-            if *negative {
-                checks.push(
-                    quote! {
-                        if self.#field_ident >= 0 {
-                            return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                                format!("Field '{}' must be negative", stringify!(#field_ident))
-                            ))
-                        }
-                    }
-                );
-            }
-        }
-
-        if let Some(non_negative) = non_negative {
-            if *non_negative {
-                checks.push(
-                    quote! {
-                        if self.#field_ident < 0 {
-                            return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                                format!("Field '{}' must be non-negative", stringify!(#field_ident))
-                            ))
-                        }
-                    }
-                );
-            }
-        }
-
-        if let Some(min) = min {
-            checks.push(
-                quote! {
-                    if (self.#field_ident as i64) < #min {
-                        return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                            format!("Field '{}' must be at least {}", stringify!(#field_ident), #min)
-                        ));
-                    }
-                }
-            );
-        }
-
-        if let Some(max) = max {
-            checks.push(
-                quote! {
-                    if (self.#field_ident as i64) > #max {
-                        return Err(::oximod::_error::oximod_error::OximodError::ValidationError(
-                            format!("Field '{}' must be at most {}", stringify!(#field_ident), #max)
-                        ));
-                    }
-                }
-            );
-        }
-
-        checks
-    });
+    let validations = validate_definitions
+        .iter()
+        .flat_map(|validate_def| generate_validate_model_tokens(validate_def));
 
     let default_inits = default_definitions.iter().map(|def| {
         let ident = &def.field_ident;
@@ -409,48 +153,8 @@ pub fn derive_model(input: TokenStream) -> TokenStream {
             quote! { #ident: Default::default(), }
         });
 
-    if has_id_attr {
-        let mut id_setter_name = "id".to_string();
-
-        for attr in &input.attrs {
-            if attr.path().is_ident("document_id_setter_ident") {
-                let setter_lit: LitStr = attr
-                    .parse_args()
-                    .expect("Expected #[document_id_setter_ident = \"...\"]");
-                id_setter_name = setter_lit.value();
-            }
-        }
-
-        let id_method_ident = syn::Ident::new(&id_setter_name, proc_macro2::Span::call_site());
-        let id_setter =
-            quote! {
-                /// Set the MongoDB ObjectId
-                pub fn #id_method_ident(mut self, id: ::oximod::_mongodb::bson::oid::ObjectId) -> Self {
-                    self._id = Some(id);
-                    self
-                }
-            };
-        setters.push(id_setter);
-    }
-
-    for (ident, ty) in all_fields.iter().filter(|(ident, _)| ident != "_id") {
-        let setter = if let Some(inner) = option_inner_type(ty) {
-            quote! {
-                    pub fn #ident<T: Into<#inner>>(mut self, val: T) -> Self {
-                        self.#ident = Some(val.into());
-                        self
-                    }
-                }
-        } else {
-            quote! {
-                    pub fn #ident(mut self, val: #ty) -> Self {
-                        self.#ident = val;
-                        self
-                    }
-                }
-        };
-        setters.push(setter);
-    }
+    maybe_push_id_setter(has_id_attr, &input.attrs, &mut setters);
+    push_field_setters(&all_fields, &mut setters);
 
     let expanded =
         quote! {
